requirements.txt
python-telegram-bot==20.8
motor==3.3.2
dnspython==2.6.1
python-dotenv==1.0.1
